rename: Modal 컴포넌트 ui폴더로 이동
